+++ packages/hydra-e2e-tests/test/e2e/api/processor-api.ts
@@ -142,9 +142,11 @@ export function subscribeToProcessorStatus(): void {
     .subscribe({
       next({ data }: unknown) {
         if (data) {
-          processorStatus = (data as {
-            stateSubscription: ProcessorStatus
-          }).stateSubscription
+          processorStatus = (
+            data as {
+              stateSubscription: ProcessorStatus
+            }
+          ).stateSubscription
         }
       },
     })