+++ packages/hydra-e2e-tests/test/e2e/transfer-e2e.test.ts
@@ -82,12 +82,8 @@ describe('end-to-end transfer tests', () => {
   })
 
   it('fetch datetime field from transfer', async () => {
-    const {
-      insertedAt,
-      createdAt,
-      updatedAt,
-      timestamp,
-    } = await fetchDateTimeFieldFromTransfer()
+    const { insertedAt, createdAt, updatedAt, timestamp } =
+      await fetchDateTimeFieldFromTransfer()
     const ts = Number.parseInt(timestamp)
 
     console.log(`Timestamp: ${timestamp}, ts: ${ts}`)